pyproject.toml
[project]
name = "docctl"
version = "0.1.0"
description = "CLI-first local document retrieval tool"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "chromadb>=0.5.23",
  "llama-index-core>=0.11.23",
  "pdfplumber",
  "pypdf",
  "reportlab",
  "sentence-transformers>=3.3.1",
  "typer>=0.12.5",
]

[project.scripts]
docctl = "docctl.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.3.4",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra"
